moving packages to pkg directory
