+++ cmd/main.go
@@ -1,6 +1,8 @@
 package main
 
-import "github.com/tamiat/backend/handlers"
+import (
+	"github.com/tamiat/backend/pkg/handlers"
+)
 
 func main(){
 	handlers.Start()
