+++ pkg/service/contentService.go
@@ -1,7 +1,7 @@
 package service
 
 import (
-	"github.com/tamiat/backend/domain/content"
+	"github.com/tamiat/backend/pkg/domain/content"
 )
 
 type ContentService interface {
@@ -20,6 +20,6 @@ func (s DefaultContentService) GetAllContents()([]content.Content,error){
 func (s DefaultContentService) GetContent(id string)(*content.Content,error){
 	return s.repo.ById(id)
 }
-func NewContentService(repository content.ContentRepository) DefaultContentService{
+func NewContentService(repository content.ContentRepository) DefaultContentService {
 	return DefaultContentService{repo: repository}
 }
